+++ src/test/java/org/tibennetwork/iamame/TestCommandLineArguments.java
@@ -3,6 +3,8 @@ package org.tibennetwork.iamame;
 import static org.hamcrest.CoreMatchers.hasItems;
 import static org.hamcrest.CoreMatchers.not;
 import static org.junit.Assert.assertThat;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertFalse;
 
 import java.io.FileInputStream;
 import java.io.IOException;
@@ -12,7 +14,6 @@ import java.util.Arrays;
 import java.util.List;
 
 import org.junit.Test;
-import org.tibennetwork.iamame.CommandLineArguments;
 import org.tibennetwork.iamame.mame.CommandLineOptions;
 import org.tibennetwork.iamame.mame.CommandLineOptionsFactory;
 import org.tibennetwork.iamame.mame.FakeMameRuntime;
@@ -22,13 +23,14 @@ import org.tibennetwork.iamame.mame.MameExecutionException;
 public class TestCommandLineArguments {
 
     @Test 
-    public void testGetRawOptionsArgsDoesntReturnMediaArgument () 
+    public void testGetMameOptionsRawArgsDoesntReturnMediaArgument () 
             throws InvalidMameArgumentsException, 
                    IOException,
                    InterruptedException,
                    MameExecutionException {
         
         // Init Options           
+
         FakeMameRuntime mame = new FakeMameRuntime();
         List<InputStream> inputStreams = new ArrayList<>();
         inputStreams.add(
@@ -45,11 +47,138 @@ public class TestCommandLineArguments {
 
         ma.validate();
 
-        String[] optionsArgs = ma.getRawOptionsArgs();
+        String[] optionsArgs = ma.getMameOptionsRawArgs();
     
         assertThat(
                 Arrays.asList(optionsArgs), 
                 not(hasItems("-cart", "dkongca")));
+
+    }
+
+    @Test 
+    public void testGetMameOptionsRawArgsDoesntReturnIaMameOptionArgument () 
+            throws InvalidMameArgumentsException, 
+                   IOException,
+                   InterruptedException,
+                   MameExecutionException {
+        
+        // Init Options           
+
+        FakeMameRuntime mame = new FakeMameRuntime();
+        List<InputStream> inputStreams = new ArrayList<>();
+        inputStreams.add(
+            new FileInputStream("src/test/resources/showconfig.txt"));
+        inputStreams.add(
+            new FileInputStream("src/test/resources/showusage.txt"));
+        mame.setInputStreamsToReturn(inputStreams);
+        CommandLineOptionsFactory clof = new CommandLineOptionsFactory ();
+        CommandLineOptions mameOpts 
+            = clof.deduceFromMameRuntime(mame);
+        
+        String[] args = {"-dryrun", "-rompath", "/tmp/roms", "punisher"};
+        CommandLineArguments ma = new CommandLineArguments(mameOpts, args);
+
+        ma.validate();
+
+        String[] optionsArgs = ma.getMameOptionsRawArgs();
+
+        assertThat(Arrays.asList(optionsArgs), not(hasItems("-dryrun")));
+
+    }
+
+    @Test 
+    public void testGetMameRawArgsDoesntReturnIaMameOptionArgument ()
+            throws InvalidMameArgumentsException, 
+                   IOException,
+                   InterruptedException,
+                   MameExecutionException {
+        
+        // Init Options           
+
+        FakeMameRuntime mame = new FakeMameRuntime();
+        List<InputStream> inputStreams = new ArrayList<>();
+        inputStreams.add(
+            new FileInputStream("src/test/resources/showconfig.txt"));
+        inputStreams.add(
+            new FileInputStream("src/test/resources/showusage.txt"));
+        mame.setInputStreamsToReturn(inputStreams);
+        CommandLineOptionsFactory clof = new CommandLineOptionsFactory ();
+        CommandLineOptions mameOpts 
+            = clof.deduceFromMameRuntime(mame);
+        
+        String[] args = {"-dryrun", "-rompath", "/tmp/roms", "punisher"};
+        CommandLineArguments ma = new CommandLineArguments(mameOpts, args);
+
+        ma.validate();
+
+        String[] optionsArgs = ma.getMameRawArgs();
+
+        assertThat(Arrays.asList(optionsArgs), not(hasItems("-dryrun")));
+
+    }
+
+    @Test 
+    public void testGetMameRawArgsDoesReturnIaMameNonOptionArgumentAndMediaTypes ()
+            throws InvalidMameArgumentsException, 
+                   IOException,
+                   InterruptedException,
+                   MameExecutionException {
+        
+        // Init Options           
+
+        FakeMameRuntime mame = new FakeMameRuntime();
+        List<InputStream> inputStreams = new ArrayList<>();
+        inputStreams.add(
+            new FileInputStream("src/test/resources/showconfig.txt"));
+        inputStreams.add(
+            new FileInputStream("src/test/resources/showusage.txt"));
+        mame.setInputStreamsToReturn(inputStreams);
+        CommandLineOptionsFactory clof = new CommandLineOptionsFactory ();
+        CommandLineOptions mameOpts 
+            = clof.deduceFromMameRuntime(mame);
+        
+        String[] args = {"-dryrun", "-rompath", "/tmp/roms", "genesis", "-cart", "sonic"};
+        CommandLineArguments ma = new CommandLineArguments(mameOpts, args);
+
+        ma.validate();
+
+        String[] optionsArgs = ma.getMameRawArgs();
+
+        assertThat(Arrays.asList(optionsArgs), not(hasItems("-dryrun")));
+        assertThat(Arrays.asList(optionsArgs), hasItems("-rompath", "/tmp/roms", 
+            "genesis", "-cart", "sonic"));
+
+    }
+
+    @Test 
+    public void testContains () 
+            throws InvalidMameArgumentsException, 
+                   IOException,
+                   InterruptedException,
+                   MameExecutionException {
+        
+        // Init Options           
+
+        FakeMameRuntime mame = new FakeMameRuntime();
+        List<InputStream> inputStreams = new ArrayList<>();
+        inputStreams.add(
+            new FileInputStream("src/test/resources/showconfig.txt"));
+        inputStreams.add(
+            new FileInputStream("src/test/resources/showusage.txt"));
+        mame.setInputStreamsToReturn(inputStreams);
+        CommandLineOptionsFactory clof = new CommandLineOptionsFactory ();
+        CommandLineOptions mameOpts 
+            = clof.deduceFromMameRuntime(mame);
+        
+        String[] args = {"-dryrun", "-rompath", "/tmp/roms", "punisher"};
+        CommandLineArguments ma = new CommandLineArguments(mameOpts, args);
+
+        ma.validate();
+
+        assertTrue(ma.contains("dryrun"));
+        assertTrue(ma.contains("rompath"));
+        assertFalse(ma.contains("punisher"));
+        assertFalse(ma.contains("waitvsync"));
 
     }